pubmed-paper-fetcher/pyproject.toml
[tool.poetry]
name = "get-papers"
version = "0.1.0"
description = "CLI tool to fetch PubMed papers with pharma/biotech authors"
authors = ["Shrubi <shrubikumari2001@gmail.com>"]
readme = "README.md"
packages = [{ include = "get_papers", from = "src" }]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "*"
biopython = "*"
requests = "^2.32.4"

[tool.poetry.scripts]
get-papers-list = "get_papers.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"







